+++ Cargo.toml
@@ -119,6 +119,7 @@ bincode = { version = "2.0.0-rc.3", features = [
 # The following packages disable the "std" feature for no_std compatibility
 #
 derive-new = { version = "0.7.0", default-features = false }
+cfg-if = "1.0.0"
 
 blas-src = { version = "0.10.0", default-features = false }
 half = { version = "2.4.1", features = [
