+++ crates/burn-core/Cargo.toml
@@ -39,6 +39,8 @@ doc = [
     "hip-jit",
     "vision",
     "autodiff",
+    "remote",
+    "server",
     # Doc features
     "burn-candle/doc",
     "burn-common/doc",
@@ -86,7 +88,8 @@ metal = ["burn-candle?/metal"]
 openblas = ["burn-ndarray?/blas-openblas"]
 openblas-system = ["burn-ndarray?/blas-openblas-system"]
 template = ["burn-wgpu?/template"]
-remote = ["burn-router-remote"]
+remote = ["burn-remote/client"]
+server = ["burn-remote/server"]
 
 candle = ["burn-candle"]
 candle-cuda = ["candle", "burn-candle/cuda"]
@@ -132,9 +135,7 @@ burn-hip = { path = "../burn-hip", version = "0.16.0", optional = true, default-
 burn-ndarray = { path = "../burn-ndarray", version = "0.16.0", optional = true, default-features = false }
 burn-tch = { path = "../burn-tch", version = "0.16.0", optional = true }
 burn-wgpu = { path = "../burn-wgpu", version = "0.16.0", optional = true, default-features = false }
-burn-router-remote = { path = "../burn-router-remote", version = "0.16.0", default-features = false, features = [
-    "client",
-], optional = true }
+burn-remote = { path = "../burn-remote", version = "0.16.0", default-features = false, optional = true }
 
 data-encoding = { workspace = true }
 uuid = { workspace = true }
