+++ crates/burn-remote/src/client/base.rs
