+++ crates/burn-remote/src/server/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod processor;
+pub(crate) mod session;
 pub(crate) mod stream;
 
 mod base;
 
-pub use base::*;
+pub use base::start;
