+++ crates/burn-remote/src/shared/mod.rs
@@ -1,0 +1,2 @@
+mod task;
+pub(crate) use task::*;
