--- crates/burn-router-remote/src/shared/mod.rs
@@ -1,3 +1,0 @@
-mod task;
-
-pub use task::*;
