+++ examples/server/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+authors = ["nathanielsimard <nathaniel.simard.42@gmail.com>"]
+edition.workspace = true
+license.workspace = true
+name = "server"
+publish = false
+version.workspace = true
+
+[features]
+cuda-jit = ["burn/cuda-jit"]
+wgpu = ["burn/wgpu"]
+ndarray = ["burn/ndarray"]
+
+[dependencies]
+cfg-if = { workspace = true }
+burn = { path = "../../crates/burn", version = "0.16.0", features = ["server"] }
