+++ examples/server/examples/server.rs
@@ -1,0 +1,3 @@
+fn main() {
+    server::start();
+}
